+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-0.2.0.0: [2018.07.22]
+0.2.0.0: [2018.07.23]
 ---------------------
 * Fixed the `Semiring` instances of `Set`, `HashSet`, `Vector`, `Storable Vector`, `Unboxed Vector`.
 * Removed the `Semiring` instances of `Seq`, `Alt`, `Endo`.
 * Added comprehensive test suite that tests all `Semiring` instances defined in Data.Semiring
-* Added Free semiring
+* Added Free semiring (Data.Semiring.Free)
 * Added newtypes: `Add`, `Mul`
+* Bounds for containers: [0.3,0.6] -> [0.5.4,0.6.0.9]
+* Add semiring instance for `Proxy`
 
 0.1.2: [2018.05.04]
 -------------------